hieu

+++ beike/Admin/Http/Controllers/TaxClassController.php
@@ -1,13 +1,5 @@
 <?php
-/**
- * TaxClassController.php
- *
- * @copyright  2022 beikeshop.com - All Rights Reserved
- * @link       https://beikeshop.com
- * @author     Edward Yang <yangjin@guangda.work>
- * @created    2022-07-26 19:45:41
- * @modified   2022-07-26 19:45:41
- */
+
 
 namespace Beike\Admin\Http\Controllers;
 
@@ -49,8 +41,11 @@ class TaxClassController extends Controller
 
     public function destroy(Request $request, int $taxClassId)
     {
-        TaxClassRepo::deleteById($taxClassId);
 
-        return json_success(trans('common.deleted_success'));
+        if ($taxClassId != 1) {
+            TaxClassRepo::deleteById($taxClassId);
+            return json_success(trans('common.deleted_success'));
+        }
+        return json_fail('Không thể xoá');
     }
 }
